chatbot_requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
requests>=2.31.0
python-multipart>=0.0.6
jinja2>=3.1.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0